Fix bug where old tasks could silently be moved to the current edition
This became and issue when we transitioned from penta to pretalx,
because talk event IDs which were unique, are now being reused as
pretalx has no knowledge of the past editions in penta. The tool ended
up pulling volunteering tasks from old editions if there was a talk ID
collision.
This fix now assumes that talk IDs are unique within each edition,
which is true, and also makes the same fix to task IDs to avoid a
possible issue caused by talks with the same name.